cpufreq: impulse: hackjob to make impulse governor work. I deserve no real credit for the work here.  There are a lot of shitty commits out there that pretend to be legit.  This isn't one of them.
I found a kernel repo with a version of impulse that actually functions well for the 3.4 kernel on msm8974.
Unfortunately there was no legit commit history.  It was clearly a kang, so I'm not going to bother mentioning the repo in my commit message.
Suffice it to say, impulse governor works on the M8 now, but I claim no credit for actually getting any of this to work.

If someone can present me with good commit history that gets me to the same place, I'd be more than happy to rebase my source to correct the problem.

Change-Id: I8e9f45451e49b8ed1825d5991aae97fba933846d